Fix 180 degree bug
This bug occured due to multithreading. Normally you aren't able to head
in the opposite direction (as you would crash into your own body)
instantly, but since input is refreshed 4-5 times for every frame, it
was possible to spam inputs to circumnavigate the check and crash into
yourself. This fixes that by a another DIRECTION variable that is only
changed every frame, so that the issue does not occur